Fix out-of-date tabnr value in WindowInfo
Closing a tabpage before the tabpage in which the WindowInfo's window is
open will cause its tabnr to decrement; caching a tabnr at construction
doesn't reflect this. Replace the tabnr member variable with a tabnr()
function that will always return an up-to-date value.

Replace an explicit v:exception with typevim#Rethrow(), to allow
standard Vim exceptions to propagate back upward.