Add Goto() func to WindowInfo; bufnr() as function
Similar to 36ab91217df2ccd2f74263d3928f22e975ff0e18, it's not enough to
set the bufnr at construction and always retrieve that cached value. Add
a function to retrieve the current bufnr value.

Use {} notation for function name declarations, for better conciseness.

Add a Goto() function, for switching to a window.