Account for weird setting variable types in tests
nvim and vim differ in whether they return a string or a number when
queried for the value of a setting like &foldcolumn. This breaks
vader.vim's AssertEqual comparisons, which checks for type equivalence.

Populate oracle dictionaries with a string or number, based on the type
of the corresponding setting.